requirements.txt
# Python 依赖包清单
# 材料管理系统 v1.8
# 生成日期：2026-03-22

# Django 核心框架
Django==6.0.3

# Django Admin 增强
django-admin-interface==0.32.0
django-colorfield==0.14.0

# Excel 文件处理
openpyxl==3.1.5

# 图像处理
Pillow==12.1.1

# PostgreSQL 数据库驱动
psycopg[binary]==3.2.6

# 环境变量管理
python-dotenv==1.2.2

# Web 服务器网关接口
gunicorn==25.1.0

# ASGI/WSGI 工具
asgiref==3.11.1

# XML 处理
et_xmlfile==2.0.0

# 打包工具
packaging==26.0

# Slug 生成
python-slugify==8.0.4
text-unidecode==1.3

# SQL 解析
sqlparse==0.5.5

# 性能监控（开发环境）
django-debug-toolbar==4.3.0

# 错误追踪（生产环境）
sentry-sdk==2.56.0

# 测试工具
pytest-xdist==3.8.0

# API框架
djangorestframework==3.17.1
django-filter==25.2

# 静态资源压缩
django-compressor==4.6.0
rcssmin==1.2.2

# Redis 缓存
django-redis==6.0.0

# 异步任务处理
celery==5.6.3
redis==7.4.0

# 测试框架
pytest-django==4.12.0
